Client:  inventory-management: Can find a stock-management with date input
